Spark 3.2: Expose actual action classes in SparkActions
